NotUsingButDontDeleteYet/requirements.txt
Flask==2.3.3
Flask-Session==0.4.0
Flask-Login==0.6.3
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.0
Celery==5.3.0
Werkzeug==3.0.4
python-dotenv==1.0.0
Flask-WTF==1.2.1
markupsafe==3.0.1
email-validator==1.3.1
alembic==1.13.3
amqp==5.2.0
billiard==4.2.1
cachelib==0.13.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.7.0
idna==3.10
kombu==5.4.2
Mako==1.3.6
prompt_toolkit==3.0.48
python-dateutil==2.9.0.post0
six==1.16.0
SQLAlchemy==1.4.43
typing_extensions==4.12.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
WTForms==3.2.1
redis==5.1.1
